requirements.txt
aniso8601==8.0.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
protobuf==4.21.2
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2021.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==2.0.15
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.3.3
WTForms==3.0.1
pandas==1.3.5
python-decouple==3.8